% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flextags}
\alias{flextags}
\title{flexTags from github.com/greymd/ojichat/pattern}
\format{
An object of class \code{list} of length 4.
}
\source{
\url{https://github.com/greymd/ojichat}
}
\usage{
flextags
}
\description{
flexTags from github.com/greymd/ojichat/pattern
}
\keyword{datasets}
