% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.R
\name{get_golem_config}
\alias{get_golem_config}
\title{Read app config}
\usage{
get_golem_config(
  value,
  config = Sys.getenv("R_CONFIG_ACTIVE", "default"),
  use_parent = TRUE
)
}
\arguments{
\item{value}{Value to retrieve from the config file.}

\item{config}{R_CONFIG_ACTIVE value.}

\item{use_parent}{Logical, scan the parent directory for config file.}
}
\description{
Read app config
}
\keyword{internal}
