% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gimei}
\alias{gimei}
\title{Fake Japanese first names ported from github.com/willnet/gimei}
\format{
An object of class \code{data.frame} with 10702 rows and 3 columns.
}
\source{
\url{https://github.com/willnet/gimei}
}
\usage{
gimei
}
\description{
Fake Japanese first names ported from github.com/willnet/gimei
}
\keyword{datasets}
