% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaketoba-package.R
\docType{package}
\name{shaketoba-package}
\alias{shaketoba}
\alias{shaketoba-package}
\title{shaketoba: R Port of Ojisan NArikiri Randomized Algorithm}
\description{
'Shaketoba', also pronounced as 'Saketoba', is salty dried autumn salmon
    that is known as a traditional food in the Hokkaido and Tohoku regions of Japan.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paithiov909/shaketoba}
  \item Report bugs at \url{https://github.com/paithiov909/shaketoba/issues}
}

}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

Other contributors:
\itemize{
  \item Guangchuang Yu [copyright holder]
}

}
\keyword{internal}
