% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_generator.R
\name{sktb_conv_ojichat}
\alias{sktb_conv_ojichat}
\title{Convert message templates into ojichat}
\usage{
sktb_conv_ojichat(
  message,
  target_name = "",
  pos_after = c("助詞", "助動詞"),
  emoji_rep = 4L,
  comma_freq = runif(1, 0, 1)
)
}
\arguments{
\item{message}{message template}

\item{target_name}{passed to \code{sktb_conv_tags_all}.}

\item{pos_after}{passed to \code{sktb_insert_comma}.}

\item{comma_freq}{passed to \code{sktb_insert_comma}.}
}
\value{
character vector.
}
\description{
Convert message templates into ojichat
}
