% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main_utils_pattern-tags.R
\name{sktb_conv_tags_all}
\alias{sktb_conv_tags_all}
\title{Convert tags in the message template}
\usage{
sktb_conv_tags_all(message, target_name = "", emoji_rep = 1L)
}
\arguments{
\item{message}{character vector.}

\item{target_name}{character vector.}

\item{emoji_rep}{integer}
}
\value{
character vector.
}
\description{
Convert tags in the message template
}
