% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_generator.R
\name{sktb_conv_tail_charclass}
\alias{sktb_conv_tail_charclass}
\title{Convert charclass of the last nth characters}
\usage{
sktb_conv_tail_charclass(message, tail = 1L, to = c("katakana", "hiragana"))
}
\arguments{
\item{message}{character vector.}

\item{tail}{integer.}

\item{to}{this argument is passed to \code{zipangu::str_conv_hirakana()}.}
}
\value{
list of character.
}
\description{
Convert charclass of the last nth characters
}
