% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_generator.R
\name{sktb_insert_comma}
\alias{sktb_insert_comma}
\title{Insert comma after specified pos}
\usage{
sktb_insert_comma(message, pos_after = c("ADP", "AUX"), comma_freq = 0)
}
\arguments{
\item{message}{character scalar.}

\item{pos_after}{character vector.}

\item{comma_freq}{numeric (if less than 0, the function does nothing).}
}
\value{
character vector.
}
\description{
Insert comma after specified pos
}
