% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{sktb_meme}
\alias{sktb_meme}
\title{meme}
\usage{
sktb_meme(
  img,
  upper = NULL,
  lower = NULL,
  size = "auto",
  color = "snow",
  font = "SetoFont-SP",
  vjust = 0.4,
  bgcolor = "violet",
  r = 0.2,
  density = NULL
)
}
\arguments{
\item{img}{path or url}

\item{upper}{upper text}

\item{lower}{lower text}

\item{size}{size of text}

\item{color}{color of text}

\item{font}{font family of text}

\item{vjust}{vertical adjustment of captions}

\item{bgcolor}{background color of shadow text}

\item{r}{ratio of shadow text}

\item{density}{resolution to render pdf or svg}
}
\value{
grob object
}
\description{
create meme
}
\examples{
f <- system.file("angry8.jpg", package="meme")
meme(f, "code", "all the things!", font = "Helvetica")
}
\author{
guangchuang yu
}
