% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pattern-onara.R
\name{sktb_message_pattern}
\alias{sktb_message_pattern}
\title{Onara message patterns}
\usage{
sktb_message_pattern(merge)
}
\arguments{
\item{merge}{list.}
}
\value{
list.
}
\description{
Onara message patterns
}
\keyword{internal}
