% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main_utils_generator.R
\name{sktb_ojichat}
\alias{sktb_ojichat}
\title{main logic ~> main server}
\usage{
sktb_ojichat(
  target_name = "",
  pos_after = c("助詞", "助動詞"),
  emoji_rep_max = 4L,
  comma_freq = runif(1, 0, 1)
)
}
\description{
main logic ~> main server
}
