% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main_utils_pattern-tags.R
\name{sktb_sample_first_name}
\alias{sktb_sample_first_name}
\title{Sample Japanese female first names}
\usage{
sktb_sample_first_name(size = 1L)
}
\arguments{
\item{size}{integer.}
}
\value{
character vector.
}
\description{
Sample Japanese female first names
}
