% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_generator.R
\name{sktb_sample_message}
\alias{sktb_sample_message}
\title{Sample message template}
\usage{
sktb_sample_message(
  size = 1L,
  merge = list(),
  tail = sample.int(3, 1),
  collapse = ""
)
}
\arguments{
\item{size}{integer.}

\item{merge}{list. internaly passed to \code{sktb_message_pattern}.}

\item{tail}{integer. passed to \code{sktb_conv_tail_charclass}.}

\item{collapse}{character scalar.}
}
\value{
message templates are returned as a charcter vector.
}
\description{
Sample message template
}
