% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main_utils_pattern-tags.R
\name{sktb_sample_name_suffix}
\alias{sktb_sample_name_suffix}
\title{Sample Japanese name suffix}
\usage{
sktb_sample_name_suffix(size = 1L)
}
\arguments{
\item{size}{integer.}
}
\value{
character vector.
}
\description{
Sample Japanese name suffix
}
