% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{sktb_udpipe}
\alias{sktb_udpipe}
\title{Wrapper of udpipe::udpipe}
\usage{
sktb_udpipe(message)
}
\arguments{
\item{message}{character vector.}
}
\value{
data.frame
}
\description{
Wrapper of udpipe::udpipe
}
\keyword{internal}
