% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{templates}
\alias{templates}
\title{OnaraMessages templates from github.com/greymd/ojichat/pattern}
\format{
An object of class \code{list} of length 8.
}
\source{
\url{https://github.com/greymd/ojichat}
}
\usage{
templates
}
\description{
OnaraMessages templates from github.com/greymd/ojichat/pattern
}
\keyword{datasets}
