% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uniqtags}
\alias{uniqtags}
\title{uniqTags from github.com/greymd/ojichat/pattern}
\format{
An object of class \code{list} of length 10.
}
\source{
\url{https://github.com/greymd/ojichat}
}
\usage{
uniqtags
}
\description{
uniqTags from github.com/greymd/ojichat/pattern
}
\keyword{datasets}
