% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{sktb_plot_ojichat}
\alias{sktb_plot_ojichat}
\title{Plot an ojichat meme}
\usage{
sktb_plot_ojichat(str, img = NULL, split = " ")
}
\arguments{
\item{str}{ojichat message.}

\item{img}{`NULL` or the file path to an image
(intenally passed to \code{magick::image_read()}).}

\item{split}{character scalar passed to \code{stringi::stri_replace_all_fixed()}
(NOT reprex).}
}
\value{
an ojichat class object is returned.
}
\description{
Plot an ojichat meme
}
