% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{sktb_plot_ojichat}
\alias{sktb_plot_ojichat}
\title{Plot an ojichat meme}
\usage{
sktb_plot_ojichat(str, img = NULL, split = "EOS")
}
\arguments{
\item{str}{ojichat message.}

\item{img}{passed to \code{meme::meme()}.}

\item{split}{character scalar.}
}
\value{
object returned from \code{meme::meme}.
}
\description{
Plot an ojichat meme
}
