% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main_utils_generator.R
\name{sktb_sample_message}
\alias{sktb_sample_message}
\title{Sample message template}
\usage{
sktb_sample_message(merge = list())
}
\arguments{
\item{merge}{list}
}
\value{
message template is returned randomly.
}
\description{
Sample message template
}
