% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main_utils_generator.R
\name{sktb_sample_message}
\alias{sktb_sample_message}
\title{Sample message template}
\usage{
sktb_sample_message(size = 1L, merge = list(), collapse = " ")
}
\arguments{
\item{size}{integer.}

\item{merge}{list.}

\item{collapse}{character scalar.}
}
\value{
message templates are returned as a charcter vector.
}
\description{
Sample message template
}
